requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
redis==5.0.1
pyyaml==6.0.1


